requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-environ==0.9.0
django-request-logging==0.7.5
djangorestframework==3.14.0
idna==3.4
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
uuid==1.30
zxcvbn==4.4.28
